avatar handling with local and remote avatar arrays as well as default avatar, user assignment is done with ARTags
