+++ .gitignore
@@ -1,0 +1,1 @@
+capturyconfig_sample.json.meta
