+++ Plugins.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 1f12ec5ee50357f4e98f9aa09004784a
+folderAsset: yes
+timeCreated: 1487783975
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
