+++ Plugins/RemoteCaptury.dll.meta
@@ -1,0 +1,21 @@
+fileFormatVersion: 2
+guid: 0d20df226e3d80044bee7fe7223758eb
+timeCreated: 1487783938
+licenseType: Free
+PluginImporter:
+  serializedVersion: 1
+  iconMap: {}
+  executionOrder: {}
+  isPreloaded: 0
+  isOverridable: 0
+  platformData:
+    Any:
+      enabled: 1
+      settings: {}
+    Editor:
+      enabled: 0
+      settings:
+        DefaultValueInitialized: true
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
