+++ README.md.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 922337918209063489629c6bb773907c
+timeCreated: 1487779940
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
