+++ Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a23e7436a6a903546922fa9270ce4dfe
+folderAsset: yes
+timeCreated: 1487782701
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
