+++ Scripts/CapturyAvatarManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1a8216c56ebc16f4b89230b4c2ae988e
+timeCreated: 1487779940
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
