+++ Scripts/CapturyConfig.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a6fb14f814d8f3342b587f1082022b45
+timeCreated: 1487936537
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
