+++ Scripts/CapturyOrigin.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fcd5c268a215705458dabfea3be81146
+timeCreated: 1487839118
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
