+++ Scripts/CapturyOriginManager.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4025b407eecedf14396a003794199483
+timeCreated: 1487860317
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
