+++ Scripts/TransformFollower.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 289abb8d72e69b04caf65a8e815c6a27
+timeCreated: 1487779940
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
