+++ capturyconfig_sample.json
@@ -1,0 +1,4 @@
+{
+	"arTagIDs" : [0,1],
+	"avatarID" : [0]
+}
